refactor(review): move the AI call into FindingPipeline.run() (#250)
Phase 2 step 4: FindingPipeline gains run(session, promptInputs, ctx,
lineResolver), which calls aiReviewService.review() and then the existing
refine() chain. aiReviewService is injected into FindingPipeline and
leaves ReviewOrchestrator — the orchestrator no longer touches the model
client, just delegates the whole finding phase. The now-unused diff local
is dropped from review().

Behaviour unchanged (same call, same chain, same order). Tests build the
pipeline with the mocked aiReviewService; integration coverage of the AI
path is via review(). Full suite green, spotbugs + spotless clean.
Orchestrator: 415 → 402 LOC, 16 → 15 collaborators.
